+++ updater/configupdate.py
@@ -7,3 +7,9 @@
 
 versioninfo = open("versioninfo.txt")
 versioninfo = versioninfo.read()
+
+if versioninfo == "0.6 beta":
+    print("No changes to your config file is needed!",
+          "Press enter to exit.", sep="\n")
+    input()
+    sys.exit()